refactor(tui): extract spine argv parsing to spine-args.ts
Addresses audit finding 35: the asOf and magnitude-threshold parsers
move verbatim out of event-store.ts into a dedicated spine-args.ts
module. spine.ts, startup.ts, and report.ts — a fourth consumer the
audit didn't name — re-point their imports to the new module. The
moved parser tests relocate to spine-args.test.ts, and parseAsOfArg
gains its first direct unit cases.
